Moved ReviewsList to features folder
