+++ src/pages/Home.tsx
@@ -1,6 +1,6 @@
 import type React from "react";
 
-import ReviewsList from "../components/ReviewsList";
+import ReviewsList from "../features/ReviewsList";
 import { reviews } from "../reviews";
 
 const Home: React.FC = () => {
